Refactor func. 'writeToOutput' to restart after '='
After the equals button has been pressed once for
some calculation, the next digit button that is pressed
will clear the result from the variable 'expression'
and begin a new calculation with the digit pressed
as the first operand. This makes the user experience
of the calculator more enjoyable to use.